requirements.txt
streamlit>=1.53.0
streamlit-js-eval>=1.0.0
pandas>=2.3.0
twilio>=9.10.0
python-dotenv>=1.0.0
bcrypt>=4.1.3
psycopg2-binary>=2.9.9
